Add files via upload
Today, we delve into the intricate world of bird strikes in the United States, focusing specifically on the years 2000 to 2011. A bird strike occurs when an aircraft collides with a bird or flock of birds during flight or on the ground. These incidents, though relatively rare, can lead to catastrophic consequences, ranging from minor damage to aircraft to catastrophic accidents resulting in loss of life