Add files via upload
Heart disease, also known as cardiovascular disease, includes conditions that affect the heart. It is the leading cause of death for both men and women. The term "heart disease" is often used interchangeably with "cardiovascular disease" and "coronary heart disease," and encompasses a range of conditions such as arrhythmias, coronary artery disease, and heart defects present at birth. Prevention, timely diagnosis, and effective management are key in reducing the impact of heart disease.